+++ rapier/src/body.rs
@@ -3,11 +3,13 @@ use bevy_math::prelude::*;
 use bevy_transform::prelude::*;
 use fnv::FnvHashMap;
 
-use heron_core::{Body, Restitution, Velocity};
+use heron_core::{Body, BodyType, Restitution, Velocity};
 
 use crate::convert::{IntoBevy, IntoRapier};
-use crate::rapier::dynamics::{JointSet, RigidBodyBuilder, RigidBodyHandle, RigidBodySet};
-use crate::rapier::geometry::{ColliderBuilder, ColliderSet};
+use crate::rapier::dynamics::{
+    BodyStatus, JointSet, RigidBodyBuilder, RigidBodyHandle, RigidBodySet,
+};
+use crate::rapier::geometry::{Collider, ColliderBuilder, ColliderSet};
 use crate::BodyHandle;
 
 pub(crate) type HandleMap = FnvHashMap<Entity, RigidBodyHandle>;
@@ -24,14 +26,17 @@ pub(crate) fn create(
             Entity,
             &Body,
             &GlobalTransform,
+            Option<&BodyType>,
             Option<&Velocity>,
             Option<&Restitution>,
         ),
         Without<BodyHandle>,
     >,
 ) {
-    for (entity, body, transform, velocity, restitution) in query.iter() {
-        let mut builder = RigidBodyBuilder::new_dynamic()
+    for (entity, body, transform, body_type, velocity, restitution) in query.iter() {
+        let body_type = body_type.cloned().unwrap_or_default();
+
+        let mut builder = RigidBodyBuilder::new(body_status(body_type))
             .user_data(entity.to_bits().into())
             .position((transform.translation, transform.rotation).into_rapier());
 
@@ -50,10 +55,7 @@ pub(crate) fn create(
 
         let rigid_body = bodies.insert(builder.build());
         let collider = colliders.insert(
-            collider_builder(&body)
-                .user_data(entity.to_bits().into())
-                .restitution(restitution.map_or(0.0, |it| (*it).into()))
-                .build(),
+            build_collider(entity, &body, body_type, restitution.cloned()),
             rigid_body,
             &mut bodies,
         );
@@ -68,18 +70,56 @@ pub(crate) fn create(
     }
 }
 
+#[allow(clippy::type_complexity)]
 pub(crate) fn update_shape(
     mut bodies: ResMut<'_, RigidBodySet>,
     mut colliders: ResMut<'_, ColliderSet>,
-    mut query: Query<'_, (&Body, &mut BodyHandle), Mutated<Body>>,
+    mut query: Query<
+        '_,
+        (
+            Entity,
+            &Body,
+            &mut BodyHandle,
+            Option<&BodyType>,
+            Option<&Restitution>,
+        ),
+        Or<(Mutated<Body>, Changed<BodyType>)>,
+    >,
 ) {
-    for (body_def, mut handle) in query.iter_mut() {
+    for (entity, body_def, mut handle, body_type, restitution) in query.iter_mut() {
         colliders.remove(handle.collider, &mut bodies, true);
         handle.collider = colliders.insert(
-            collider_builder(&body_def).build(),
+            build_collider(
+                entity,
+                &body_def,
+                body_type.cloned().unwrap_or_default(),
+                restitution.cloned(),
+            ),
             handle.rigid_body,
             &mut bodies,
         );
+    }
+}
+
+pub(crate) fn update_rapier_status(
+    mut bodies: ResMut<'_, RigidBodySet>,
+    with_type_changed: Query<'_, (&BodyType, &BodyHandle), Changed<BodyType>>,
+    without_type: Query<'_, &BodyHandle, Without<BodyType>>,
+) {
+    for (body_type, handle) in with_type_changed.iter() {
+        if let Some(body) = bodies.get_mut(handle.rigid_body) {
+            body.body_status = body_status(*body_type);
+        }
+    }
+
+    for entity in without_type.removed::<BodyType>() {
+        if let Some(body) = without_type
+            .get(*entity)
+            .ok()
+            .and_then(|handle| bodies.get_mut(handle.rigid_body))
+        {
+            body.body_status = body_status(BodyType::default());
+        }
     }
 }
 
@@ -99,10 +139,22 @@ pub(crate) fn update_rapier_position(
 
 pub(crate) fn update_bevy_transform(
     bodies: Res<'_, RigidBodySet>,
-    mut query: Query<'_, (Option<&mut Transform>, &mut GlobalTransform, &BodyHandle)>,
+    mut query: Query<
+        '_,
+        (
+            Option<&mut Transform>,
+            &mut GlobalTransform,
+            &BodyHandle,
+            Option<&BodyType>,
+        ),
+    >,
 ) {
-    for (mut local, mut global, handle) in query.iter_mut() {
-        let body = match bodies.get(handle.rigid_body).filter(|it| it.is_dynamic()) {
+    for (mut local, mut global, handle, body_type) in query.iter_mut() {
+        if !matches!(body_type.cloned().unwrap_or_default(), BodyType::Dynamic) {
+            continue;
+        }
+
+        let body = match bodies.get(handle.rigid_body) {
             None => continue,
             Some(body) => body,
         };
@@ -148,15 +200,30 @@ pub(crate) fn remove(
     }
 }
 
-fn collider_builder(body: &Body) -> ColliderBuilder {
-    match body {
+fn build_collider(
+    entity: Entity,
+    body: &Body,
+    body_type: BodyType,
+    restitution: Option<Restitution>,
+) -> Collider {
+    let mut builder = match body {
         Body::Sphere { radius } => ColliderBuilder::ball(*radius),
         Body::Capsule {
             half_segment: half_height,
             radius,
         } => ColliderBuilder::capsule_y(*half_height, *radius),
         Body::Cuboid { half_extends } => cuboid_builder(*half_extends),
+    };
+
+    builder = builder
+        .user_data(entity.to_bits().into())
+        .sensor(matches!(body_type, BodyType::Sensor));
+
+    if let Some(restitution) = restitution {
+        builder = builder.restitution(restitution.into());
     }
+
+    builder.build()
 }
 
 #[inline]
@@ -171,16 +238,29 @@ fn cuboid_builder(half_extends: Vec3) -> ColliderBuilder {
     ColliderBuilder::cuboid(half_extends.x, half_extends.y, half_extends.z)
 }
 
+fn body_status(body_type: BodyType) -> BodyStatus {
+    match body_type {
+        BodyType::Dynamic => BodyStatus::Dynamic,
+        BodyType::Static | BodyType::Sensor => BodyStatus::Static,
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use super::*;
     use bevy_math::Vec3;
+
+    use super::*;
 
     #[test]
     fn build_sphere() {
-        let builder = collider_builder(&Body::Sphere { radius: 4.2 });
+        let builder = build_collider(
+            Entity::new(0),
+            &Body::Sphere { radius: 4.2 },
+            BodyType::default(),
+            None,
+        );
         let ball = builder
-            .shape
+            .shape()
             .as_ball()
             .expect("Created shape was not a ball");
         assert_eq!(ball.radius, 4.2);
@@ -188,11 +268,16 @@ mod tests {
 
     #[test]
     fn build_cuboid() {
-        let builder = collider_builder(&Body::Cuboid {
-            half_extends: Vec3::new(1.0, 2.0, 3.0),
-        });
+        let builder = build_collider(
+            Entity::new(0),
+            &Body::Cuboid {
+                half_extends: Vec3::new(1.0, 2.0, 3.0),
+            },
+            BodyType::default(),
+            None,
+        );
         let cuboid = builder
-            .shape
+            .shape()
             .as_cuboid()
             .expect("Created shape was not a cuboid");
 
@@ -205,12 +290,17 @@ mod tests {
 
     #[test]
     fn build_capsule() {
-        let builder = collider_builder(&Body::Capsule {
-            half_segment: 10.0,
-            radius: 5.0,
-        });
+        let builder = build_collider(
+            Entity::new(0),
+            &Body::Capsule {
+                half_segment: 10.0,
+                radius: 5.0,
+            },
+            BodyType::default(),
+            None,
+        );
         let capsule = builder
-            .shape
+            .shape()
             .as_capsule()
             .expect("Created shape was not a capsule");